refactor(material): Handout-Mechanik in eigene Datei + MaterialHandoutShell
Re-Audit nach Tier-1-Abschluss (#107, #112, #113): MaterialPageTemplate
mischte Page-Chrome (Hero, FAQ-Cluster, Block-Header) mit Handout-Mechanik
(3 Helper + 3 Sub-Components + 3 Handout-Komponenten + Switch). Vor der
Tier-2-Welle (#H-3, #H-4, #H-7) wird die Mechanik herausgezogen, damit die
Page-Datei nicht weiter aufgeblasen wird.

Was sich aendert:
- Neue Datei src/templates/MaterialHandouts.jsx (~440 Zeilen) buendelt
  alle handout-bezogenen Komponenten + den Default-Export
  MaterialHandoutSwitch.
- Neuer Wrapper MaterialHandoutShell eliminiert die wiederholte
  Article+Head+Usage+...+Disclaimer+CrossRefs-Boilerplate aus
  MaterialConversationScript + MaterialThresholdChecklist. Die DOM-Struktur
  bleibt bit-identisch -- nur die JSX-Wiederholung verschwindet.
- MaterialPageTemplate.jsx schrumpft von 590 auf 184 Zeilen und enthaelt
  nur noch Page-Level-Rendering. Der Handout-Switch wird per Default-Import
  eingebunden.

Was sich NICHT aendert:
- MaterialCrisisPlan bleibt inline in MaterialHandouts.jsx und nutzt
  bewusst weder Shell noch geteilte Helper -- DOM-Stabilitaet fuer die
  bestehenden Print-Tests + Print-Typografie in app-global.css §9.
  Konsolidierung kommt erst bei begruendetem Anlass (z. B. Label-
  Vereinheitlichung wie "Wie lange gilt er/es").
- Alle BEM-Klassen, IDs, data-handout-id-Attribute, ARIA-Labels bleiben
  identisch -- die 3 Material-Handout-Tests sehen keine Veraenderung.

CLAUDE.md aktualisiert: Verweis auf neue Datei, MaterialHandoutShell als
empfohlener Wrapper fuer neue Handouts dokumentiert.

Verifikation:
- npm run lint: clean
- npx prettier --check: clean
- npm run build: clean (Bundle 57.08 -> 57.02 kB, marginal kleiner)
- npx playwright test: 37/37 gruen, inkl. aller 3 Material-Handout-Tests
